+++ test/spec/actions/change_preset.js
@@ -165,15 +165,132 @@ describe('iD.actionChangePreset', function() {
         });
     });
 
-    // https://github.com/openstreetmap/iD/issues/9372
-    it('does not preserve field tags when changing from a subpreset to its parent', function() {
-        var entity = new iD.osmNode({tags: {highway: 'service', service: 'driveway'}});
-        var graph = new iD.coreGraph([entity]);
-        var oldPreset = iD.presetPreset('highway/service/driveway', {tags: {highway: 'service', service: 'driveway'}});
-        var newPreset = iD.presetPreset('highway/service', {tags: {highway: 'service'}, fields: ['field']}, undefined, {
-            field: iD.presetField('field', {key: 'service'})
+    // https://github.com/openstreetmap/iD/issues/12075
+    it.each([{
+        fieldType: 'multiCombo',
+        fieldId: 'recycling',
+        fieldKey: 'recycling:',
+        fieldKeys: [],
+        oldTags: { 'recycling:paper': 'yes', 'recycling:others': 'no' },
+        tagsToPreserve: {}
+    }, {
+        fieldType: 'localized',
+        fieldId: 'name',
+        fieldKey: 'name',
+        fieldKeys: [],
+        oldTags: { 'name': 'foo', 'name:en': 'bar', 'name:etymology:wikidata': 'Q860' },
+        tagsToPreserve: { 'name:etymology:wikidata': 'Q860' }
+    }, {
+        fieldType: 'directionalCombo',
+        fieldId: 'cycleway',
+        fieldKey: 'cycleway',
+        fieldKeys: [ 'cycleway:left', 'cycleway:right' ],
+        oldTags: { 'cycleway:left': 'no', 'cycleway:both': 'separate', 'cycleway:foo': 'bar' },
+        tagsToPreserve: { 'cycleway:foo': 'bar' }
+    }])('does not preserve $fieldType field tags that are only present in the old preset', ({
+        fieldType, fieldId, fieldKey, fieldKeys, oldTags, tagsToPreserve
+    }) => {
+        const entity = iD.osmNode({
+            tags: {
+                amenity: 'recycling',
+                ...oldTags,
+                unrelatedTagKey: 'unrelatedTagValue',
+            },
+            loc: [0, 0],
         });
-        var action = iD.actionChangePreset(entity.id, oldPreset, newPreset);
-        expect(action(graph).entity(entity.id).tags).to.eql({highway: 'service'});
+        const graph = new iD.coreGraph([entity]);
+
+        const fields = {
+            [fieldId]: iD.presetField('recycling', { type: fieldType, key: fieldKey, keys: fieldKeys }),
+        };
+
+        const oldPreset = iD.presetPreset(
+            'amenity/recycling',
+            { tags: { amenity: 'recycling' }, fields: [fieldId] },
+            undefined,
+            fields,
+        );
+        const newPreset = iD.presetPreset(
+            'amenity/bench',
+            { tags: { amenity: 'bench' }, fields: [] },
+            undefined,
+            fields,
+        );
+        const action = iD.actionChangePreset(entity.id, oldPreset, newPreset);
+        expect(action(graph).entity(entity.id).tags).toStrictEqual({
+            // no field tags are preserved
+            amenity: 'bench',
+            ...tagsToPreserve,
+            unrelatedTagKey: 'unrelatedTagValue',
+        });
+    });
+
+    // https://github.com/openstreetmap/iD/pull/12218#issuecomment-4314204446
+    it.each([{
+        fieldType: 'multiCombo',
+        fieldId: 'recycling',
+        fieldKey: 'recycling:',
+        fieldKeys: [],
+        oldTags: { 'recycling:paper': 'yes', 'recycling:others': 'no' }
+    }, {
+        fieldType: 'localized',
+        fieldId: 'name',
+        fieldKey: 'name',
+        fieldKeys: [],
+        oldTags: { 'name': 'foo', 'name:en': 'bar', 'name:etymology:wikidata': 'Q860' }
+    }, {
+        fieldType: 'directionalCombo',
+        fieldId: 'cycleway',
+        fieldKey: 'cycleway',
+        fieldKeys: [ 'cycleway:left', 'cycleway:right' ],
+        oldTags: { 'cycleway:left': 'no', 'cycleway:both': 'separate', 'cycleway:foo': 'bar' }
+    }])('preserve $fieldType field tags when they are present in the old and the new preset', ({
+        fieldType, fieldId, fieldKey, fieldKeys, oldTags
+    }) => {
+        const entity = iD.osmNode({
+            tags: {
+                amenity: 'recycling',
+                ...oldTags,
+            },
+            loc: [0, 0],
+        });
+        const graph = new iD.coreGraph([entity]);
+
+        const fields = {
+            [fieldId]: iD.presetField('recycling', { type: fieldType, key: fieldKey, keys: fieldKeys }),
+        };
+
+        const oldPreset = iD.presetPreset(
+            'amenity/recycling',
+            { tags: { amenity: 'recycling' }, fields: [fieldId] },
+            undefined,
+            fields,
+        );
+        const newPreset = iD.presetPreset(
+            'amenity/bench',
+            { tags: { amenity: 'bench' }, fields: [fieldId] },
+            undefined,
+            fields,
+        );
+        const action = iD.actionChangePreset(entity.id, oldPreset, newPreset);
+        expect(action(graph).entity(entity.id).tags).toStrictEqual({
+            // all field tags are preserved
+            ...oldTags,
+            amenity: 'bench', // override primary preset tag
+        });
+    });
+
+    // https://github.com/openstreetmap/iD/issues/9372
+    it('does not preserve old preset\'s primary tags when changing from a subpreset to its parent', () => {
+        const entity = new iD.osmNode({tags: {highway: 'service', service: 'driveway', name: 'foo bar'}});
+        const graph = new iD.coreGraph([entity]);
+        const fields = {
+            field: iD.presetField('field', {key: 'service'}),
+            name: iD.presetField('name', {key: 'name'})
+        };
+        const oldPreset = iD.presetPreset('highway/service/driveway', {tags: {highway: 'service', service: 'driveway'}, fields: ['name']}, undefined, fields);
+        const newPreset = iD.presetPreset('highway/service', {tags: {highway: 'service'}, fields: ['field', 'name']}, undefined, fields);
+        const action = iD.actionChangePreset(entity.id, oldPreset, newPreset);
+        expect(action(graph).entity(entity.id).tags).to.eql({highway: 'service', name: 'foo bar'});
     });
 });